pyproject.toml
[project]
name = "spelling-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "edge-tts>=6.1.19",
    "llama-index-llms-ollama>=0.3.6",
]

